🎨 split socketService into multiple services and add interfaces
