--- src/app/district.ts
@@ -1,4 +1,0 @@
-export interface District {
-  cost: number;
-  placeOnBoard?: THREE.Vector3;
-}
