+++ src/app/interfaces/board-element.ts
@@ -1,0 +1,3 @@
+export interface BoardElement {
+  placeOnBoard?: THREE.Vector3;
+}
