+++ src/app/interfaces/character.ts
@@ -1,0 +1,5 @@
+export interface Character {
+  dead:boolean;
+  role: string;
+  target?: string;
+}
