+++ src/app/interfaces/district.ts
@@ -1,0 +1,7 @@
+import { BoardElement } from "./board-element";
+
+export interface District extends BoardElement{
+  cost: number;
+  disctrictType?: string;
+  name?: string;
+}
