--- src/app/player.ts
@@ -1,8 +1,0 @@
-import { District } from './district';
-
-export interface Player {
-  id: string;
-  name: string;
-  citadel: District[];
-  positionOnBoard?: THREE.Vector3;
-}
